chore: refactor paypal configuration to enhance modularity
